rename: 이미지 파일들 public 폴더로 이동
