+++ components/AllArticleSection.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import formatDate from "@/util/formatDate";
-import Heart from "@/assets/images/icons/ic_heart.svg";
-import Profile from "@/assets/images/icons/ic_profile.svg";
+import Heart from "@/public/images/icons/ic_heart.svg"
+import Profile from "@/public/images/icons/ic_profile.svg";
 import styles from "@/components/AllArticleSection.module.css";
 import DropDownButton from "./DropDownButton";
 import SearchForm from "./SearchForm";
