+++ components/DropDownButton.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Sort from "@/assets/images/icons/ic_sort.svg";
+import Sort from "@/public/images/icons/ic_sort.svg";
 import styles from "@/components/DropDownButton.module.css";
 import { useState } from "react";
 import { useRouter } from "next/router";
