+++ components/FileInput.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import Image from "next/image";
-import Close from "@/assets/images/icons/ic_x.svg";
+import Close from "@/public/images/icons/ic_x.svg";
 import styles from "@/components/FileInput.module.css";
 interface Props {
   name: string;
