+++ components/ProductCard.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import Heart from "@/assets/images/icons/ic_heart.svg";
+import Heart from "@/public/images/icons/ic_heart.svg";
 import styles from "@/components/ProductCard.module.css";
 import { Product } from "@/types/type";
