+++ components/SearchForm.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Search from "@/assets/images/icons/ic_search.svg";
+import Search from "@/public/images/icons/ic_search.svg";
 import styles from "@/components/SearchForm.module.css";
 
 interface SearchFormProps {
