+++ pages/login.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
 import Image from "next/image";
-import Logo from "@/assets/images/logo/logo.svg";
-import Google from "@/assets/images/social/google-logo.png";
-import Kakao from "@/assets/images/social/kakao-logo.png";
+import Logo from "@/public/images/logo/logo.svg";
+import Google from "@/public/images/social/google-logo.png";
+import Kakao from "@/public/images/social/kakao-logo.png";
 import styles from "@/styles/SignPage.module.css";
 import React, { useEffect, useState } from "react";
