+++ public/images/home/hero-image.png
